Rename the application
Using `app` as the name of the application is confusing since we have an
`app` variable as well.
