+++ cgap/__init__.py
@@ -28,7 +28,7 @@ def create_app(test_config=None):
 		def hello():
 				return 'Hello, World!'
 
-		from . import db
+		from cgap import db
 		db.init_app(app)
 
 		return app
